social_oculus/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
dataclasses-json==0.5.9
dnspython==2.3.0
exceptiongroup==1.1.2
fastapi==0.100.0
grpcio==1.30.0
grpcio-tools==1.30.0
h11==0.14.0
idna==3.4
marshmallow==3.19.0
marshmallow-enum==1.5.1
minio==7.1.15
motor==3.2.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pip==22.3.1
protobuf==3.20.0
pydantic==1.10.9
pydantic_core==2.1.2
pymongo==4.4.0
python-dotenv==1.0.0
python-facebook-api==0.15.0
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
typing-inspect==0.9.0
urllib3==2.0.3
uvicorn==0.22.0
wheel==0.38.4
colorlog==6.7.0
